fix(babel): fixed babel6 and babel7
