+++ plugins/mrbuilder-plugin-babel-6/bin/cli.js
@@ -1,12 +1,10 @@
 #!/usr/bin/env node
 if (!global._MRBUILDER_OPTIONS_MANAGER) {
-    process.env.MRBUILDER_INTERNAL_PLUGINS=`${process.env.MRBUILDER_INTERNAL_PLUGINS ||''},mrbuilder-plugin-babel`;
-    global._MRBUILDER_OPTIONS_MANAGER =
-        new (require('mrbuilder-optionsmanager').default)
+    process.env.MRBUILDER_INTERNAL_PLUGINS = `${process.env.MRBUILDER_INTERNAL_PLUGINS || ''},mrbuilder-plugin-babel,mrbuilder-plugin-babel-6`;
+    global._MRBUILDER_OPTIONS_MANAGER      = new (require('mrbuilder-optionsmanager').default)
 }
-const { plugins, presets }        = require(
-    'mrbuilder-plugin-babel/babel-config');
-const { argv }                    = process;
+const {plugins, presets} = require('../babel-config');
+const {argv}             = process;
 if (!argv.includes('-s')) {
     argv.push('-s', 'true');
 }
@@ -25,4 +23,4 @@ if (!argv.includes('--plugins')) {
 if (!(argv.includes('--out-file') || argv.includes('--filename'))) {
     argv.push('src');
 }
-require('@babel/cli/bin/babel');
+require('babel-cli/bin/babel');
