--- plugins/mrbuilder-plugin-babel-7/bin/clean-cli.js
@@ -1,3 +1,0 @@
-#!/usr/bin/env node
-process.argv.push('lib', "dist");
-require('rimraf');
