+++ plugins/mrbuilder-plugin-babel-7/bin/cli.js
@@ -1,12 +1,18 @@
 #!/usr/bin/env node
-if (!global._MRBUILDER_OPTIONS_MANAGER) {
-    process.env.MRBUILDER_INTERNAL_PLUGINS=`${process.env.MRBUILDER_INTERNAL_PLUGINS ||''},mrbuilder-plugin-babel`;
-    global._MRBUILDER_OPTIONS_MANAGER =
-        new (require('mrbuilder-optionsmanager').default)
+const tmp     = require('tmp');
+const fs      = require('fs');
+
+let om = global._MRBUILDER_OPTIONS_MANAGER;
+if (!om) {
+    process.env.MRBUILDER_INTERNAL_PLUGINS = `${process.env.MRBUILDER_INTERNAL_PLUGINS || ''},mrbuilder-plugin-babel,mrbuilder-plugin-babel-7`;
+
+    om = global._MRBUILDER_OPTIONS_MANAGER = new (require('mrbuilder-optionsmanager').default)
 }
-const { plugins, presets }        = require(
-    'mrbuilder-plugin-babel/babel-config');
-const { argv }                    = process;
+
+const {argv} = process;
+
+let useConfig = true;
+
 if (!argv.includes('-s')) {
     argv.push('-s', 'true');
 }
@@ -16,13 +22,21 @@ if (!(argv.includes('--out-file') || argv.includes('--out-dir'))) {
 if (!(argv.includes('--copy-files') || argv.includes('-D'))) {
     argv.push('--copy-files');
 }
-if (!argv.includes('--presets')) {
-    argv.push('--presets', presets);
+if (argv.includes('--presets')) {
+    useConfig = false;
 }
-if (!argv.includes('--plugins')) {
-    argv.push('--plugins', plugins);
+if (argv.includes('--plugins')) {
+    useConfig = false
 }
 if (!(argv.includes('--out-file') || argv.includes('--filename'))) {
     argv.push('src');
 }
-require('babel-cli/bin/babel');
+if (argv.includes('--config-file')) {
+    useConfig = false;
+}
+let tmpobj = tmp.fileSync();
+if (useConfig) {
+    fs.writeFileSync(tmpobj.name, JSON.stringify(require(`../babel-config`), null, 2), 'utf8');
+    argv.push('--config-file', tmpobj.name);
+}
+require('@babel/cli/bin/babel');
